Include cuda instance memory cuh

+++ src/inst/cuda_instance.cu
@@ -16,7 +16,7 @@ void gs::cuda::inst::copy_to_symbol(const instance64<uint32_t, uint32_t>& inst)
 }
 
 __device__  bool gs::cuda::inst::has_connection_to(const uint64_t* adjacency, uint32_t from, uint32_t to) {
-	if (adjacency[from] & (1 << to)) return true;
+	if (adjacency[from] & (uint64_t(1) << to)) return true;
 	else return false;
 }
